// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/linking-resolved-source-page-dto-api-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { LinkingResolvedSourcePageDto } from '../models/linking-resolved-source-page-dto';
export interface LinkingResolvedSourcePageDtoApiResponse {
  data: LinkingResolvedSourcePageDto | null;
  errors: Array<string> | null;
  isSuccess: boolean;
  message: string | null;
}